+++ Docs/CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.6.670] - 2025-07-17
+## [0.6.680] - 2025-07-18
 ### Fixed
 - Fixed online chat update in OpenTgResearcherDesktop
+- Fixed chats page in OpenTgResearcherDesktop
+- Fixed client connection page in OpenTgResearcherDesktop
 ### Changed
 - Updated message table in TgStorage
 - Updated the main page in OpenTgResearcherDesktop
@@ -20,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added chat details view in OpenTgResearcherDesktop
 - Added viewing of chat participants in OpenTgResearcherDesktop
 - Added viewing of chat statistics in OpenTgResearcherDesktop
-- 
+
 ## [0.6.600] - 2025-07-12
 ### Fixed
 - Fixed storage update in OpenTgResearcherDesktop